requirements.txt
fastapi
gtts
python-dotenv
uvicorn
azure-storage-blob